requirements.txt
flask==3.0.0
openai-whisper==20231117
google-generativeai==0.8.3
python-dotenv==1.0.0
